+++ src/views/Report.vue
@@ -258,7 +258,6 @@
     },
     async mounted() {
 
-
       const reportData = localStorage.getItem('report-data');
       if (reportData !== null) {
         this.reportData = JSON.parse(reportData);
@@ -266,7 +265,6 @@
 
       if (this.reportData.sessionId === null) {
         this.reportData.sessionId = uuidv4();
-        localStorage.setItem('session-id', this.sessionId);
       }
 
       if (this.reportData.diagnostic !== null) {
@@ -349,7 +347,6 @@
         }
 
         try {
-          this.reportData.lastReport = new Date();
 
           await this.$recaptchaLoaded();
 
@@ -370,7 +367,6 @@
               sessionId: this.reportData.sessionId,
               symptoms: symptoms,
               diagnostic: this.reportData.diagnostic,
-              timestamp: this.reportData.lastReport,
               appVersion: process.env.VERSION,
             }),
           });
@@ -378,6 +374,8 @@
           if (!response.ok) {
             throw new Error('could not report');
           }
+
+          this.reportData.lastReport = new Date();
 
           localStorage.setItem('report-data', JSON.stringify(this.reportData));
           this.forceReportAgain = false;